Refactor timetable scroll offset lofic.
